+++ src/JSONs.js
@@ -11,19 +11,47 @@ var util = require('util');
 
 var InvalidValueError = require('./InvalidValueError');
 
-// TODO: add support for JSON.stringify's "replacer" parameter
-// TODO: add detection of recursive references (see json-stringify-safe)
-
 /**
- * @param {*} data
+ * @param {Object} data
  * @param {Array.<(string|number)>} references
  * @param {?function(string,*)} replacer
+ * @return {undefined}
+ */
+function checkObject(data, references, replacer) {
+    if (typeof data.toJSON === 'function') {
+        return check(data.toJSON(), references, replacer);
+    }
+
+    for (var key in data) { // jshint ignore:line
+        var value = replacer ? replacer(key, data[key]) : data[key];
+
+        if (!(replacer && value === undefined)) {
+            check(value, references.concat(key), replacer);
+        }
+    }
+}
+
+/**
+ * @param {Array} data
+ * @param {Array.<(string|number)>} references
+ * @param {?function(string,*)} replacer
+ * @return {undefined}
+ */
+function checkArray(data, references, replacer) {
+    return data.forEach(function (item, index) {
+        var value = replacer ? replacer(index, item) : item;
+        check(value, references.concat(index), replacer);
+    });
+}
+
+/**
+ * Give more helpful error messages for most common invalid types.
+ *
+ * @param {*} data
+ * @param {Array.<(string|number)>} references
  * @throws {JSONs.InvalidValueError}
  */
-function check(data, references, replacer) {
-    /*jshint maxcomplexity:false*/
-
-    // Give more helpful error messages for the most common errors
+function checkCommonTypes(data, references) {
     if (util.isError(data)) {
         throw new InvalidValueError('Error is not a valid JSON type', data, references);
     }
@@ -39,44 +67,39 @@ function check(data, references, replacer) {
     if (typeof data === 'number' && !isFinite(data)) {
         throw new InvalidValueError('non-finite number is not a valid JSON type', data, references);
     }
+}
 
-    // Primitive types are always okay
+/**
+ * @param {*} data
+ * @param {Array.<(string|number)>} references
+ * @param {?function(string,*)} replacer
+ * @throws {JSONs.InvalidValueError}
+ */
+function check(data, references, replacer) {
+    checkCommonTypes(data, references);
+
+    // Primitive types are always okay (we've already checked for non-finite numbers)
     if (data === null || typeof data === 'string' || typeof data === 'boolean' || typeof data === 'number') {
         return;
     }
 
     // If an array, check its elements
     if (Array.isArray(data)) {
-        data.forEach(function (item, index) {
-            var value = replacer ? replacer(index, item) : item;
-            check(value, references.concat(index), replacer);
-        });
-        return;
+        return checkArray(data, references, replacer);
     }
 
+    /* istanbul ignore else */
     // If an object, check its properties
-    var value;
     if (typeof data === 'object') {
-        if (typeof data.toJSON === 'function') {
-            check(data.toJSON(), references, replacer);
-        } else {
-            for (var key in data) { // jshint ignore:line
-                value = replacer ? replacer(key, data[key]) : data[key];
-
-                if (!(replacer && value === undefined)) {
-                    check(value, references.concat(key), replacer);
-                }
-            }
-        }
-        return;
+        return checkObject(data, references, replacer);
     }
 
-    // Anything else is an error
+    /* istanbul ignore next */
+    // Anything else (e.g., a host object) is an error
     throw new InvalidValueError('invalid type', data, references);
 }
 
 /**
- *
  * @param {(function(string,*)|Array.<(string|number)>)=} replacer
  * @return {?function(string,*)}
  */
@@ -87,11 +110,7 @@ function normalizeReplacer(replacer) {
         };
     }
 
-    if (typeof replacer === 'function') {
-        return replacer;
-    }
-
-    return null;
+    return typeof replacer === 'function' ? replacer : null;
 }
 
 /**