+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.2.0 (2014-??-??)
-
  - added support for JSON.stringify's `replacer` parameter
- - improved error message
+ - added support for `toJSON` methods
+ - improved error messages
 
 ## 0.1.0 (2014-09-18)